+++ AppCenter/AppCenter/Internals/Model/CommonSchema/MSCSModelConstants.h
@@ -21,9 +21,9 @@ extern NSString *const kMSCSVerValue;
 
 extern NSString *const kMSUserLocale;
 
-#pragma mark - MSLocExtension
+#pragma mark - MSDeviceExtension
 
-extern NSString *const kMSTimezone;
+extern NSString *const kMSDeviceLocalId;
 
 #pragma mark - MSOsExtension
 
@@ -54,6 +54,10 @@ extern NSString *const kMSSDKEpoch;
 extern NSString *const kMSSDKSeq;
 extern NSString *const kMSSDKInstallId;
 
+#pragma mark - MSLocExtension
+
+extern NSString *const kMSTimezone;
+
 #pragma mark - MSCSData
 
 extern NSString *const kMSDataBaseDataType;
@@ -63,6 +67,7 @@ extern NSString *const kMSDataBaseData;
 
 extern NSString *const kMSCSProtocolExt;
 extern NSString *const kMSCSUserExt;
+extern NSString *const kMSCSDeviceExt;
 extern NSString *const kMSCSOSExt;
 extern NSString *const kMSCSAppExt;
 extern NSString *const kMSCSNetExt;