+++ AppCenterAnalytics/AppCenterAnalytics/TransmissionTarget/MSPropertyConfigurator.h
@@ -50,6 +50,17 @@ NS_ASSUME_NONNULL_BEGIN
  */
 - (void)removeEventPropertyForKey:(NSString *)propertyKey;
 
+/**
+ * Once called, the App Center SDK will automatically add
+ * UIDevice.identifierForVendor to common schema logs.
+ *
+ * @discussion Call this before starting the SDK. This setting is not persisted,
+ * so you need to call this when setting up the SDK every time. If you want to
+ * provide a way for users to opt-in or opt-out of this setting, it is on you to
+ * persist their choice and configure the App Center SDK accordingly.
+ */
+- (void)collectDeviceId;
+
 NS_ASSUME_NONNULL_END
 
 @end