+++ AppCenterAnalytics/AppCenterAnalytics/TransmissionTarget/MSPropertyConfigurator.m
@@ -1,12 +1,24 @@
-#import "MSPropertyConfigurator.h"
+#import "MSPropertyConfiguratorPrivate.h"
+
+#if TARGET_OS_OSX
+#import <IOKit/IOKitLib.h>
+#else
+#import <UIKit/UIKit.h>
+#endif
+
 #import "MSAnalyticsInternal.h"
 #import "MSAnalyticsTransmissionTargetInternal.h"
 #import "MSAnalyticsTransmissionTargetPrivate.h"
 #import "MSCommonSchemaLog.h"
 #import "MSLogger.h"
-#import "MSPropertyConfiguratorPrivate.h"
 
 @implementation MSPropertyConfigurator
+
+#if TARGET_OS_OSX
+static const char deviceIdPrefix = 'u';
+#else
+static const char deviceIdPrefix = 'i';
+#endif
 
 - (instancetype)initWithTransmissionTarget:
     (MSAnalyticsTransmissionTarget *)transmissionTarget {
@@ -52,6 +64,10 @@
   }
 }
 
+- (void)collectDeviceId {
+  self.deviceId = [MSPropertyConfigurator getDeviceIdentifier];
+}
+
 #pragma mark - MSChannelDelegate
 
 - (void)channel:(id<MSChannelProtocol>)__unused channel
@@ -71,7 +87,7 @@
     // Override the application version.
     while (target) {
       if (target.propertyConfigurator.appVersion) {
-        [((MSCommonSchemaLog *)log)ext].appExt.ver =
+        ((MSCommonSchemaLog *)log).ext.appExt.ver =
             target.propertyConfigurator.appVersion;
         break;
       }
@@ -82,7 +98,7 @@
     target = self.transmissionTarget;
     while (target) {
       if (target.propertyConfigurator.appName) {
-        [((MSCommonSchemaLog *)log)ext].appExt.name =
+        ((MSCommonSchemaLog *)log).ext.appExt.name =
             target.propertyConfigurator.appName;
         break;
       }
@@ -93,13 +109,46 @@
     target = self.transmissionTarget;
     while (target) {
       if (target.propertyConfigurator.appLocale) {
-        [((MSCommonSchemaLog *)log)ext].appExt.locale =
+        ((MSCommonSchemaLog *)log).ext.appExt.locale =
             target.propertyConfigurator.appLocale;
         break;
       }
       target = target.parentTarget;
     }
+
+    // The device ID must not be inherited from parent transmission targets.
+    [((MSCommonSchemaLog *)log)ext].deviceExt.localId = self.deviceId;
   }
+}
+
+#pragma mark - Helper methods
+
++ (NSString *)getDeviceIdentifier {
+  NSString *baseIdentifier;
+#if TARGET_OS_OSX
+  /*
+   * TODO: Uncomment this for macOS support.
+   * io_service_t platformExpert = IOServiceGetMatchingService(
+   *    kIOMasterPortDefault, IOServiceMatching("IOPlatformExpertDevice"));
+   * CFStringRef platformUUIDAsCFString = NULL;
+   * if (platformExpert) {
+   *  platformUUIDAsCFString = (CFStringRef)IORegistryEntryCreateCFProperty(
+   *      platformExpert, CFSTR(kIOPlatformUUIDKey), kCFAllocatorDefault, 0);
+   *  IOObjectRelease(platformExpert);
+   * }
+   * NSString *platformUUIDAsNSString = nil;
+   * if (platformUUIDAsCFString) {
+   *   platformUUIDAsNSString =
+   *    [NSString stringWithString:(__bridge NSString *)platformUUIDAsCFString];
+   *   CFRelease(platformUUIDAsCFString);
+   * }
+   * baseIdentifier = platformUUIDAsNSString;
+   */
+  baseIdentifier = @"";
+#else
+  baseIdentifier = [[[UIDevice currentDevice] identifierForVendor] UUIDString];
+#endif
+  return [NSString stringWithFormat:@"%c:%@", deviceIdPrefix, baseIdentifier];
 }
 
 @end