+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * **[Fix]** Add missing network request error logging.
 
+### AppCenterAnalytics
+
+* **[Feature]** Preparation work for a future change in transmission protocol and endpoint for Analytics data. There is no impact on your current workflow when using App Center.
+
 ___
 
 ## Version 1.9.0